backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets>=14.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.3
